+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Custom operations registered with `addCreateWidgetOperation` can now specify an `ifTypesIntersect` property containing an array of widget type names. If the area in question allows at least one, the operation is offered.
 * The login-requirements tests were updated to include tests for the `uponSubmit` filter
 * Add `prependNodes` and `appendNodes` calls for `main`.
+* Add options for pieces to change the title, message and icon of the empty state block.
 
 ### Fixes